requirements.txt
pandas==1.3.4
tabula-py==2.3.0
PyPDF2==3.0.0
openpyxl==3.0.9
